Uploaded image of the old myIIT welcome page
The academic section is not immediately visible when you get to this page. It will be better if the registration tab is immediately available for the user when they log in.
We feel the icons are all over the place. I believe the gmail should have a better icon to display messages. Icons are meant to convey messages without reading a long text.