Uploaded image of the old myIIT course searching page
When you get to this page, there should be a section to select your major or college before the subjects for each major will be displayed instead all the subjects for all majors are put together. This is quite confusing for the user since you have to scroll through a long list to find the subject for your specific major.